+++ demo/DemoApp.EFDataExample/RulesEngineContext.cs
@@ -16,29 +16,23 @@ namespace RulesEngine.Data
         {
             base.OnModelCreating(modelBuilder);
 
-            modelBuilder.Entity<ScopedParam>()
-              .HasKey(k => k.Name);
+            //modelBuilder.Entity<RulesEngine.Models.Rule>(entity => {
+            //    entity.HasKey(k => k.Name); //adds shadow property Name1
+            //    entity.Ignore(b => b.WorkflowsToInject);
+            //    entity.Ignore(b => b.WorkflowRulesToInject); //has a get which is why this line is necessary
 
-            modelBuilder.Entity<Workflow>(entity => {
-                entity.HasKey(k => k.WorkflowName);
-                entity.Ignore(b => b.WorkflowsToInject);
-            });
+            //    entity.Property(b => b.Properties).HasConversion(
+            //        v => JsonSerializer.Serialize(v, (JsonSerializerOptions)null),
+            //        v => JsonSerializer.Deserialize<Dictionary<string, object>>(v, (JsonSerializerOptions)null),
+            //        new ValueComparer<Dictionary<string, object>>(
+            //            (c1, c2) => c1.SequenceEqual(c2),
+            //            c => c.Aggregate(0, (a, v) => HashCode.Combine(a, v.GetHashCode())),
+            //            c => c));
 
-            modelBuilder.Entity<Rule>(entity => {
-                entity.HasKey(k => k.RuleName);
-
-                entity.Property(b => b.Properties)
-                .HasConversion(
-                    v => JsonSerializer.Serialize(v, null),
-                    v => JsonSerializer.Deserialize<Dictionary<string, object>>(v, null));
-
-                entity.Property(p => p.Actions)
-                .HasConversion(
-                    v => JsonSerializer.Serialize(v, null),
-                   v => JsonSerializer.Deserialize<RuleActions>(v, null));
-
-                entity.Ignore(b => b.WorkflowsToInject);
-            });
+            //    entity.Property(p => p.Actions).HasConversion(
+            //        v => JsonSerializer.Serialize(v, (JsonSerializerOptions)null),
+            //        v => JsonSerializer.Deserialize<RulesEngine.Models.RuleActions>(v, (JsonSerializerOptions)null));                
+            //});
         }
     }